# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fleet Tracking API",
        "description": "API Documentation for Fleet Live Tracking system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://fleet.local",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user and get access token",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|xxxxxxxxxxxxxx"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user and invalidate token",
                "operationId": "79854151e609bd20032aa5aa4d5a35c0",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/telemetry": {
            "post": {
                "tags": [
                    "Telemetry"
                ],
                "summary": "Submit vehicle telemetry data",
                "operationId": "c2b1dc566602894a49d4dfc88e2f8612",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "lat",
                                    "lng",
                                    "speed",
                                    "timestamp"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -6.2
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 106.816666
                                    },
                                    "speed": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 60.5
                                    },
                                    "timestamp": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-10-08T10:30:00Z"
                                    },
                                    "snapshot": {
                                        "description": "Optional image snapshot",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data received successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/telemetry/latest": {
            "get": {
                "tags": [
                    "Telemetry"
                ],
                "summary": "Get latest fleet positions",
                "operationId": "aeb3453db48a27ca29bf3b60ef3f2e3e",
                "responses": {
                    "200": {
                        "description": "A list of the latest vehicle positions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}